requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.1
pydantic-settings==2.1.0
openai==1.12.0
python-multipart==0.0.9
aiohttp==3.9.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
python-magic==0.4.27
SpeechRecognition==3.10.1
pydub==0.25.1
numpy==1.26.3
whisper==1.0.0
ffmpeg-python==0.2.0